move to model ts
